+++ lib/rodar_bpmn/application.ex
@@ -11,6 +11,7 @@ defmodule RodarBpmn.Application do
         {Registry, keys: :duplicate, name: RodarBpmn.EventRegistry},
         RodarBpmn.Registry,
         RodarBpmn.TaskRegistry,
+        RodarBpmn.Expression.ScriptRegistry,
         {DynamicSupervisor, name: RodarBpmn.ContextSupervisor, strategy: :one_for_one},
         {DynamicSupervisor, name: RodarBpmn.ProcessSupervisor, strategy: :one_for_one},
         RodarBpmn.Event.Start.Trigger