+++ test/rodar_bpmn/activity/task/script_test.exs
@@ -98,7 +98,7 @@ defmodule RodarBpmn.Activity.Task.ScriptTest do
         {:bpmn_activity_task_script,
          %{id: "task", outgoing: ["flow_out"], type: "python", script: "print(1)"}}
 
-      assert {:error, "Unsupported script language: python. Only Elixir and FEEL are supported."} =
+      assert {:error, "Unsupported script language: \"python\""} =
                Script.token_in(elem, context)
     end
 
@@ -110,8 +110,7 @@ defmodule RodarBpmn.Activity.Task.ScriptTest do
         {:bpmn_activity_task_script,
          %{id: "task", outgoing: ["flow_out"], type: "javascript", script: "1+1"}}
 
-      assert {:error,
-              "Unsupported script language: javascript. Only Elixir and FEEL are supported."} =
+      assert {:error, "Unsupported script language: \"javascript\""} =
                Script.token_in(elem, context)
     end
   end